+++ Solgema/fullcalendar/browser/views.py
@@ -645,7 +645,7 @@ class SolgemaFullcalendarColorsCssFolder(BrowserView):
         css = ''
         if not colorsDict or not availableSubFolders:
             return css
-        folderIds = [a.getId for a in self.context.getFolderContents(contentFilter={'portal_type':'Folder'})]
+        folderIds = [a.getId for a in self.context.getFolderContents(contentFilter={'object_provides':'Products.ATContentTypes.interfaces.folder.IATFolder'})]
         if not folderIds:
             return css
         fieldid = 'subFolders'